CustomSelectControlV2: Remove legacy adapter layer
