+++ packages/components/src/custom-select-control-v2/default-component/index.tsx
@@ -8,8 +8,11 @@ import * as Ariakit from '@ariakit/react';
  */
 import _CustomSelect from '../custom-select';
 import type { CustomSelectProps } from '../types';
+import type { WordPressComponentProps } from '../../context';
 
-function CustomSelect( props: CustomSelectProps ) {
+function CustomSelect(
+	props: WordPressComponentProps< CustomSelectProps, 'button', false >
+) {
 	const { defaultValue, onChange, value, ...restProps } = props;
 	// Forward props + store from v2 implementation
 	const store = Ariakit.useSelectStore( {
