+++ packages/components/src/custom-select-control-v2/index.tsx
@@ -1,5 +1,5 @@
 /**
  * Internal dependencies
  */
-export { default as CustomSelect } from './legacy-adapter';
+export { default as CustomSelect } from './default-component';
 export { default as CustomSelectItem } from './custom-select-item';
